(#451) ♻️: PairUserUseCaseProtocol 분리
