+++ Doolda/Doolda/Domain/Interfaces/UseCases/PairUserUseCaseProtocol.swift
@@ -1,0 +1,18 @@
+//
+//  PairUserUseCaseProtocol.swift
+//  Doolda
+//
+//  Created by 정지승 on 2021/11/29.
+//
+
+import Combine
+import Foundation
+
+protocol PairUserUseCaseProtocol {
+    var pairedUserPublisher: Published<User?>.Publisher { get }
+    var errorPublisher: Published<Error?>.Publisher { get }
+    
+    func pair(user: User, friendId: DDID)
+    func pair(user: User)
+    func disconnectPair(user: User) -> AnyPublisher<User, Error>
+}
